requirements.txt
# Training
--index-url https://pypi.org/simple
--extra-index-url https://download.pytorch.org/whl/cu121
torch
torchvision
torchaudio
Pillow

# Project Specific
chess
meson
stockfish

# Basic Packages
rich
shapely
lightning
hydra-core
hydra-colorlog
hydra-optuna-sweeper
opencv-contrib-python

# Data
rasterio
pydantic
dvc-gdrive
SQLAlchemy

# Logger
wandb
mlflow
tensorboard
tensorboardX

# Visualization
matplotlib

#### Development Related ####
sh
pytest
ipykernel
ipywidgets
pre-commit
rootutils
autorootcwd
pyrootutils
